Java Fundamentals/src/com/revature/hw1/q15/OpsTest.java
package com.revature.hw1.q15;

import static org.junit.Assert.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

public class OpsTest {

	Ops op;
	
	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
		op = new Ops();
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void test() {
		assertEquals(6, op.subtraction(8, 2));
		assertEquals(10, op.multiplication(5, 2));
	}

}
